version 0.0.1

+++ keras/datasets/cifar10.py
@@ -10,7 +10,7 @@ def load_data(test_split=0.1, seed=113):
     path = get_file(dirname, origin=origin, untar=True)
 
     nb_samples = 50000
-    X = np.zeros((nb_samples, 3, 32, 32), dtype="uint8")
+    X = np.zeros((nb_samples, 3, 32, 32), dtype="float64")
     y = np.zeros((nb_samples,))
     for i in range(1, 6):
         fpath = path + '/data_batch_' + str(i)
@@ -38,4 +38,3 @@ def load_data(test_split=0.1, seed=113):
     y_test = y[int(len(X)*(1-test_split)):]
 
     return (X_train, y_train), (X_test, y_test)
-
